requirements.txt
instagrapi
Pillow

